refactor: folder 재정비
