+++ src/components/bookNote/Drawer1.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { DrawerWrapper } from "..";
+import { DrawerWrapper } from ".";
 
 export default function Drawer1() {
   return (
