+++ src/components/bookNote/index.ts
@@ -1,11 +1,11 @@
-export { default as Drawer1 } from "./common/Drawer1";
-export { default as Drawer2 } from "./common/Drawer2";
-export { default as Drawer3 } from "./common/Drawer3";
-export { default as DrawerWrapper } from "./common/DrawerWrapper";
-export { default as ExButton } from "./common/ExButton";
-export { default as StepUp } from "./common/StepUp";
+export { default as Drawer1 } from "./Drawer1";
+export { default as Drawer2 } from "./Drawer2";
+export { default as Drawer3 } from "./Drawer3";
+export { default as DrawerWrapper } from "./DrawerWrapper";
+export { default as ExButton } from "./ExButton";
 export { default as Navigator } from "./Navigator";
 export { default as PeriNote } from "./periNote/PeriNote";
 export { default as InputQuestion } from "./preNote/InputQuestion";
 export { default as PreNoteForm } from "./preNote/PreNoteForm";
 export { default as QuestionThree } from "./preNote/QuestionThree";
+export { default as StepUp } from "./StepUp";
